frontend/requirements.txt
streamlit==1.37.1
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.0
